+++ NEWS.md
@@ -1,5 +1,6 @@
 # MRFcov 1.0.40
 *   Improved nonparanormal mapping of predictions for discrete variables
+*   Option to remove progress bar in multicore estimation, significantly speeding up analysis runs
 
 # MRFcov 1.0.39
 *   Improvement for the detection of infinite values in data and coordinates